Use 64-bit type for device size so we can properly display the size of device >4GB.

+++ src/main.c
@@ -100,7 +100,7 @@ char factor_symbols[] = "kMGTPE";
  *
  * @param size	Data size to print
  */
-void print_human_size(size_t size) {
+void print_human_size(uint64_t size) {
 	float fsize = 0;
 	unsigned int factor = 0;
